Jobs will not be erased in production mode

+++ database/migrations/2019_08_19_000000_create_failed_jobs_table.php
@@ -29,6 +29,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('failed_jobs');
+        if (App::isLocal()) {
+            Schema::dropIfExists('failed_jobs');
+        }
     }
 };
