+++ database/migrations/2023_12_18_180011_create_jobs_table.php
@@ -29,6 +29,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('jobs');
+        if (App::isLocal()) {
+            Schema::dropIfExists('jobs');
+        }
     }
 };
